vision-service/requirements.txt
flask>=2.0
flask-cors>=3.0
pillow>=9.0
pillow-heif>=0.13
torch>=2.0
torchvision>=0.15
transformers>=4.30
